Move another peer_info function into the correct location
