+++ src/peer_info.h
@@ -55,6 +55,13 @@ size_t purge_expired_nodes (struct peer_info **peer_list,
                             time_t *p_last_purge,
                             int frequency, int timeout);
 
+struct peer_info* add_sn_to_list_by_mac_or_sock (
+    struct peer_info **sn_list,
+    n2n_sock_t *sock,
+    const n2n_mac_t mac,
+    int *skip_add
+);
+
 int find_and_remove_peer (struct peer_info **, const n2n_mac_t);
 struct peer_info* find_peer_by_sock (const n2n_sock_t *, struct peer_info *);
